Metodi deposito e prelievo

+++ src/main/java/com/novavrbe/vrbe/dto/InventoryDto.java
@@ -16,5 +16,5 @@ public class InventoryDto {
     @Column(name = "CHARACTER_ID")
     private Integer characterId;
     @Column
-    private BigDecimal gold;
+    private Integer gold;
 }
